+++ game folder/Assets/Scripts/SaveObjects/EquipmentData.cs
@@ -8,7 +8,7 @@ public class EquipmentData {
 
     public string m_equipmentName = "";
 	public string m_prefabName = "";
-	public int m_equipmentLevel = 1;
+	public float m_equipmentLevel = 1;
 	public float m_creditValue = 1.0f;
 	public EnergyType m_damageType = EnergyType.proton;
 	public string m_Owner = "player";
